+++ src/pco-api-client.ts
@@ -1,56 +1,16 @@
-import { ApplicationConfig } from './application/application.js';
-import { Giving } from './application/giving.js';
 import { defaultsDeep } from 'lodash-es';
+import { PartialDeep } from 'type-fest';
 import { HttpClient } from './http-client/index.js';
 import { Calendar } from './application/calendar.js';
 import { CheckIns } from './application/check-ins.js';
+import { Giving } from './application/giving.js';
 import { Groups } from './application/groups.js';
 import { People } from './application/people.js';
 import { Services } from './application/services.js';
-
-export interface PCOApiClientConfig {
-  apiVersion: string;
-  applications: PCOApiClientAppConfig;
-  auth: PCOApiClientAuthPAT; // TODO: Add support for OAuth2.
-}
-
-export interface PCOApiClientAuthPAT {
-  secret: string;
-  token: string;
-}
-
-export interface PCOApiClientAppConfig {
-  calendar: ApplicationConfig;
-  checkIns: ApplicationConfig;
-  giving: ApplicationConfig;
-  groups: ApplicationConfig;
-  people: ApplicationConfig;
-  services: ApplicationConfig;
-}
-
-export const PCOApiClientConfigDefaults: Omit<PCOApiClientConfig, 'auth'> = {
-  apiVersion: 'v2',
-  applications: {
-    calendar: {
-      version: '2021-07-20',
-    },
-    checkIns: {
-      version: '2023-04-05',
-    },
-    giving: {
-      version: '2019-10-18',
-    },
-    groups: {
-      version: '2018-08-01',
-    },
-    people: {
-      version: '2023-02-15',
-    },
-    services: {
-      version: '2018-11-01',
-    },
-  },
-};
+import {
+  PCOApiClientConfig,
+  PCOApiClientConfigDefaults,
+} from './pco-api-client-config.js';
 
 export class PCOApiClient {
   private client: HttpClient;
@@ -61,51 +21,50 @@ export class PCOApiClient {
   people: People;
   services: Services;
 
-  constructor(options: PCOApiClientConfig) {
-    const token = 'token' in options.auth ? options.auth.token : '';
-    const secret = 'secret' in options.auth ? options.auth.secret : '';
-    this.client = new HttpClient({ auth: { token, secret } });
+  constructor(options?: PartialDeep<PCOApiClientConfig>) {
+    const config: PCOApiClientConfig = defaultsDeep(
+      options,
+      PCOApiClientConfigDefaults
+    ) as PCOApiClientConfig;
 
-    defaultsDeep(options, PCOApiClientConfigDefaults);
+    this.client = new HttpClient({
+      auth: { token: config.auth.token, secret: config.auth.secret },
+    });
 
     this.calendar = new Calendar(
       this.client,
-      options.apiVersion,
-      options.applications.calendar
+      config.apiVersion,
+      config.applications.calendar
     );
 
     this.checkIns = new CheckIns(
       this.client,
-      options.apiVersion,
-      options.applications.checkIns
+      config.apiVersion,
+      config.applications.checkIns
     );
 
     this.giving = new Giving(
       this.client,
-      options.apiVersion,
-      options.applications.giving
+      config.apiVersion,
+      config.applications.giving
     );
 
     this.groups = new Groups(
       this.client,
-      options.apiVersion,
-      options.applications.groups
+      config.apiVersion,
+      config.applications.groups
     );
 
     this.people = new People(
       this.client,
-      options.apiVersion,
-      options.applications.people
+      config.apiVersion,
+      config.applications.people
     );
 
     this.services = new Services(
       this.client,
-      options.apiVersion,
-      options.applications.services
+      config.apiVersion,
+      config.applications.services
     );
-  }
-
-  setDefaults(options: PCOApiClientConfig) {
-    defaultsDeep(options, PCOApiClientConfigDefaults);
   }
 }